% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_r_files_from_github}
\alias{get_r_files_from_github}
\title{Get R file name and url from github}
\usage{
get_r_files_from_github(repo = "hadley/dplyr")
}
\arguments{
\item{repo}{Github repository address with format username/repo}
}
\value{
a list mapping R file names to github download urls
}
\description{
Get R file name and url from github
}

