% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{type_contents}
\alias{type_contents}
\title{Type contents of list of expressions}
\usage{
type_contents(expression_group)
}
\arguments{
\item{expression_group}{list of expressions}
}
\description{
Type contents of list of expressions
}
