% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{type_github}
\alias{type_github}
\title{Typing practice via R packages on Github}
\usage{
type_github(repo = "tidyverse/dplyr")
}
\arguments{
\item{repo}{Which repo do you want to practice typing on ("tidyverse/dplyr") is
the default}
}
\description{
This function sends the user into an interactive loop that simulates
typing a part of an actual R package on github. The standard
words-per-minute calculations are presented to the user after each typed
line and following the typing round.
}
\examples{
\dontrun{
type_github()
}

}
