% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-utils.R
\name{visit_site}
\alias{visit_site}
\title{Visit typingtutor's companion website}
\usage{
visit_site()
}
\description{
Uses utils::browseURL to launch a browser with destination
\url{https://baogorek.github.io/typingtutor/site/signed-in.html}
}
